+++ common/src/main/java/de/cech12/usefulhats/item/HatArmorModels.java
@@ -1,23 +1,29 @@
 package de.cech12.usefulhats.item;
 
 import de.cech12.usefulhats.Constants;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.item.equipment.EquipmentAsset;
+import net.minecraft.world.item.equipment.EquipmentAssets;
 
 public class HatArmorModels {
 
-    public static final ResourceLocation AQUANAUT = Constants.id("aquanaut_helmet");
-    public static final ResourceLocation BUNNY = Constants.id("bunny_ears");
-    public static final ResourceLocation CHOPPING = Constants.id("chopping_hat");
-    public static final ResourceLocation ENDER = Constants.id("ender_helmet");
-    public static final ResourceLocation HALO = Constants.id("halo");
-    public static final ResourceLocation LUCKY = Constants.id("lucky_hat");
-    public static final ResourceLocation MINING = Constants.id("mining_hat");
-    public static final ResourceLocation MUSHROOM = Constants.id("mushroom_hat");
-    public static final ResourceLocation POSTMAN = Constants.id("postman_hat");
-    public static final ResourceLocation SHULKER = Constants.id("shulker_helmet");
-    public static final ResourceLocation STOCKING = Constants.id("stocking_cap");
-    public static final ResourceLocation STOCKING_XMAS = Constants.id("stocking_cap_xmas");
-    public static final ResourceLocation STRAW = Constants.id("straw_hat");
-    public static final ResourceLocation WING = Constants.id("wing_helmet");
+    public static final ResourceKey<EquipmentAsset> AQUANAUT = createId("aquanaut_helmet");
+    public static final ResourceKey<EquipmentAsset> BUNNY = createId("bunny_ears");
+    public static final ResourceKey<EquipmentAsset> CHOPPING = createId("chopping_hat");
+    public static final ResourceKey<EquipmentAsset> ENDER = createId("ender_helmet");
+    public static final ResourceKey<EquipmentAsset> HALO = createId("halo");
+    public static final ResourceKey<EquipmentAsset> LUCKY = createId("lucky_hat");
+    public static final ResourceKey<EquipmentAsset> MINING = createId("mining_hat");
+    public static final ResourceKey<EquipmentAsset> MUSHROOM = createId("mushroom_hat");
+    public static final ResourceKey<EquipmentAsset> POSTMAN = createId("postman_hat");
+    public static final ResourceKey<EquipmentAsset> SHULKER = createId("shulker_helmet");
+    public static final ResourceKey<EquipmentAsset> STOCKING = createId("stocking_cap");
+    public static final ResourceKey<EquipmentAsset> STOCKING_XMAS = createId("stocking_cap_xmas");
+    public static final ResourceKey<EquipmentAsset> STRAW = createId("straw_hat");
+    public static final ResourceKey<EquipmentAsset> WING = createId("wing_helmet");
+
+    private static ResourceKey<EquipmentAsset> createId(String name) {
+        return ResourceKey.create(EquipmentAssets.ROOT_ID, Constants.id(name));
+    }
 
 }