+++ common/src/main/java/de/cech12/usefulhats/item/ShulkerHelmetItem.java
@@ -22,7 +22,7 @@ public class ShulkerHelmetItem extends AbstractHatItem implements IEquipmentChan
     private static final int LEVITATION_DURATION = 219;
 
     public ShulkerHelmetItem(String name) {
-        super(name, HatArmorMaterials.SHULKER, rawColorFromRGB(150, 105, 150), Services.CONFIG::getShulkerHelmetDurability, Services.CONFIG::isShulkerHelmetDamageEnabled);
+        super(name, HatArmorMaterials.SHULKER, Services.CONFIG::getShulkerHelmetDurability, Services.CONFIG::isShulkerHelmetDamageEnabled);
     }
 
     @Override