+++ common/src/main/java/de/cech12/usefulhats/item/StockingCapItem.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class StockingCapItem extends AbstractHatItem {
 
     public StockingCapItem(String name) {
-        super(name, HatArmorMaterials.STOCKING, rawColorFromRGB(204, 0, 23), Services.CONFIG::getStockingCapDurability, Services.CONFIG::isStockingCapDamageEnabled);
+        super(name, HatArmorMaterials.STOCKING, Services.CONFIG::getStockingCapDurability, Services.CONFIG::isStockingCapDamageEnabled);
     }
 
     /**