+++ fabric/src/main/java/de/cech12/usefulhats/client/UsefulHatsFabricClientMod.java
@@ -2,16 +2,13 @@ package de.cech12.usefulhats.client;
 
 import de.cech12.usefulhats.Constants;
 import de.cech12.usefulhats.client.compat.AccessoriesClientCompat;
-import de.cech12.usefulhats.item.AbstractHatItem;
 import de.cech12.usefulhats.platform.Services;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.rendering.v1.ArmorRenderer;
-import net.fabricmc.fabric.api.client.rendering.v1.ColorProviderRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityModelLayerRegistry;
 import net.minecraft.client.model.geom.LayerDefinitions;
 import net.minecraft.client.model.geom.ModelLayerLocation;
 import net.minecraft.client.renderer.entity.state.HumanoidRenderState;
-import net.minecraft.world.item.component.DyedItemColor;
 
 public class UsefulHatsFabricClientMod implements ClientModInitializer {
 
@@ -25,8 +22,6 @@ public class UsefulHatsFabricClientMod implements ClientModInitializer {
         EntityModelLayerRegistry.registerModelLayer(USEFUL_HAT_LAYER, () -> UsefulHatModel.createLayer(LayerDefinitions.INNER_ARMOR_DEFORMATION, 0));
         UsefulHatsArmorRenderer renderer = new UsefulHatsArmorRenderer();
         Services.REGISTRY.getAllHatItems().forEach(item -> {
-            //register item colors
-            ColorProviderRegistry.ITEM.register((itemStack, layer) -> layer > 0 ? -1 : DyedItemColor.getOrDefault(itemStack, ((AbstractHatItem)itemStack.getItem()).getDefaultColor()), item);
             //register armor renderer
             ArmorRenderer.register(renderer, item);
         });