+++ neoforge/src/main/java/de/cech12/usefulhats/client/UsefulHatItemExtension.java
@@ -3,8 +3,8 @@ package de.cech12.usefulhats.client;
 import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.model.Model;
 import net.minecraft.client.renderer.entity.state.HumanoidRenderState;
+import net.minecraft.client.resources.model.EquipmentClientInfo;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.equipment.EquipmentModel;
 import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
 import org.jetbrains.annotations.NotNull;
 
@@ -17,16 +17,11 @@ public class UsefulHatItemExtension implements IClientItemExtensions {
 
     @Override
     @NotNull
-    public Model getHumanoidArmorModel(@NotNull ItemStack itemStack, @NotNull EquipmentModel.LayerType layerType, @NotNull Model original) {
+    public Model getHumanoidArmorModel(@NotNull ItemStack itemStack, @NotNull EquipmentClientInfo.LayerType layerType, @NotNull Model original) {
         if (original instanceof HumanoidModel<?> humanoidModel) {
             ((HumanoidModel<HumanoidRenderState>) humanoidModel).copyPropertiesTo(UsefulHatsNeoForgeClientEvents.usefulHatModel);
         }
         return UsefulHatsNeoForgeClientEvents.usefulHatModel;
-    }
-
-    @Override
-    public int getDefaultDyeColor(@NotNull ItemStack stack) {
-        return AbstractUsefulHatsRenderer.getDefaultColor(stack);
     }
 
 }